requirements.txt
proxmoxer~=2.1.0
paho-mqtt~=2.1.0
bidict~=0.23.1
requests~=2.32.3
mmh3~=4.1.0